# app/src/api/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "JInshanBodyShopApi",
    "version": "1.0"
  },
  "paths": {
...
    "/Personal/Products": {
      "get": {
        "tags": ["Order"],
        "parameters": [
          {
            "name": "orderId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Company/Products": {
      "get": {
        "tags": ["Order"],
        "parameters": [
          {
            "name": "orderId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/GetOrders/Personal": {
      "get": {
        "tags": ["Order"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "orderByColumn",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderAscending",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "orderByStatus",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/OrderStatus"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalOrderApiGetReturnType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalOrderApiGetReturnType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalOrderApiGetReturnType"
                }
              }
            }
          }
        }
      }
    },
    "/GetOrders/Company": {
      "get": {
        "tags": ["Order"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "orderByColumn",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderAscending",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "orderByStatus",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/OrderStatus"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyOrderApiGetReturnType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyOrderApiGetReturnType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyOrderApiGetReturnType"
                }
              }
            }
          }
        }
      }
    },
    "/GetOrder/Personal": {
      "get": {
        "tags": ["Order"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalOrder"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalOrder"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalOrder"
                }
              }
            }
          }
        }
      }
    },
    "/GetOrder/Company": {
      "get": {
        "tags": ["Order"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyOrder"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyOrder"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyOrder"
                }
              }
            }
          }
        }
      }
    },
    "/GetOrder/Company/Status": {
      "get": {
        "tags": ["Order"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "email",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "number",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyOrder"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyOrder"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyOrder"
                }
              }
            }
          }
        }
      }
    },
    "/GetOrder/Personal/Status": {
      "get": {
        "tags": ["Order"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "email",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "number",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalOrder"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalOrder"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalOrder"
                }
              }
            }
          }
        }
      }
    },
    "/PostCompanyOrder": {
      "post": {
        "tags": ["Order"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostCompanyOrder"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostCompanyOrder"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostCompanyOrder"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
    "/PostPersonalOrder": {
      "post": {
        "tags": ["Order"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostPersonalOrder"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostPersonalOrder"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostPersonalOrder"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
    "/PutPersonalOrder": {
      "put": {
        "tags": ["Order"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/OrderStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/PutCompanyOrder": {
      "put": {
        "tags": ["Order"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/OrderStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompanyOrder": {
        "required": [
          "companyName",
          "companyNip",
          "contactNumber",
          "email",
          "orderAddress",
          "orderCity",
          "orderCountry",
          "orderPostalCode",
          "orderState",
          "orderValue",
          "shippingAddress",
          "shippingCity",
          "shippingCountry",
          "shippingName",
          "shippingOption",
          "shippingPostalCode",
          "shippingState",
          "shippingSurname"
        ],
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "contactNumber": {
            "type": "string",
            "nullable": true
          },
          "orderCountry": {
            "type": "string",
            "nullable": true
          },
          "orderState": {
            "type": "string",
            "nullable": true
          },
          "orderAddress": {
            "type": "string",
            "nullable": true
          },
          "orderCity": {
            "type": "string",
            "nullable": true
          },
          "orderPostalCode": {
            "type": "string",
            "nullable": true
          },
          "shippingCountry": {
            "type": "string",
            "nullable": true
          },
          "shippingState": {
            "type": "string",
            "nullable": true
          },
          "shippingAddress": {
            "type": "string",
            "nullable": true
          },
          "shippingCity": {
            "type": "string",
            "nullable": true
          },
          "shippingPostalCode": {
            "type": "string",
            "nullable": true
          },
          "shippingName": {
            "type": "string",
            "nullable": true
          },
          "shippingSurname": {
            "type": "string",
            "nullable": true
          },
          "shippingOption": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "number": {
            "type": "integer",
            "format": "int64"
          },
          "orderValue": {
            "type": "integer",
            "format": "int32"
          },
          "companyName": {
            "type": "string",
            "nullable": true
          },
          "companyNip": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CompanyOrderApiGetReturnType": {
        "required": ["items"],
        "type": "object",
        "properties": {
          "maxPages": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompanyOrder"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OrderStatus": {
        "enum": [0, 1, 2, 3, 4, 5],
        "type": "integer",
        "format": "int32"
      },
      "PersonalOrder": {
        "required": [
          "contactNumber",
          "email",
          "orderAddress",
          "orderCity",
          "orderCountry",
          "orderName",
          "orderPostalCode",
          "orderState",
          "orderSurname",
          "orderValue",
          "shippingAddress",
          "shippingCity",
          "shippingCountry",
          "shippingName",
          "shippingOption",
          "shippingPostalCode",
          "shippingState",
          "shippingSurname"
        ],
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "contactNumber": {
            "type": "string",
            "nullable": true
          },
          "orderCountry": {
            "type": "string",
            "nullable": true
          },
          "orderState": {
            "type": "string",
            "nullable": true
          },
          "orderAddress": {
            "type": "string",
            "nullable": true
          },
          "orderCity": {
            "type": "string",
            "nullable": true
          },
          "orderPostalCode": {
            "type": "string",
            "nullable": true
          },
          "shippingCountry": {
            "type": "string",
            "nullable": true
          },
          "shippingState": {
            "type": "string",
            "nullable": true
          },
          "shippingAddress": {
            "type": "string",
            "nullable": true
          },
          "shippingCity": {
            "type": "string",
            "nullable": true
          },
          "shippingPostalCode": {
            "type": "string",
            "nullable": true
          },
          "shippingName": {
            "type": "string",
            "nullable": true
          },
          "shippingSurname": {
            "type": "string",
            "nullable": true
          },
          "shippingOption": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "number": {
            "type": "integer",
            "format": "int64"
          },
          "orderValue": {
            "type": "integer",
            "format": "int32"
          },
          "orderName": {
            "type": "string",
            "nullable": true
          },
          "orderSurname": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PersonalOrderApiGetReturnType": {
        "required": ["items"],
        "type": "object",
        "properties": {
          "maxPages": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PersonalOrder"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PostCompanyOrder": {
        "required": [
          "companyName",
          "companyNip",
          "contactNumber",
          "email",
          "orderAddress",
          "orderCity",
          "orderCountry",
          "orderPostalCode",
          "orderProducts",
          "orderState",
          "orderValue",
          "shippingAddress",
          "shippingCity",
          "shippingCountry",
          "shippingName",
          "shippingOption",
          "shippingPostalCode",
          "shippingState",
          "shippingSurname"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "contactNumber": {
            "type": "string",
            "nullable": true
          },
          "orderCountry": {
            "type": "string",
            "nullable": true
          },
          "orderState": {
            "type": "string",
            "nullable": true
          },
          "orderAddress": {
            "type": "string",
            "nullable": true
          },
          "orderCity": {
            "type": "string",
            "nullable": true
          },
          "orderPostalCode": {
            "type": "string",
            "nullable": true
          },
          "shippingCountry": {
            "type": "string",
            "nullable": true
          },
          "shippingState": {
            "type": "string",
            "nullable": true
          },
          "shippingAddress": {
            "type": "string",
            "nullable": true
          },
          "shippingCity": {
            "type": "string",
            "nullable": true
          },
          "shippingPostalCode": {
            "type": "string",
            "nullable": true
          },
          "shippingName": {
            "type": "string",
            "nullable": true
          },
          "shippingSurname": {
            "type": "string",
            "nullable": true
          },
          "shippingOption": {
            "type": "string",
            "nullable": true
          },
          "orderValue": {
            "type": "integer",
            "format": "int32"
          },
          "orderProducts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductAndQuantity"
            },
            "nullable": true
          },
          "companyName": {
            "type": "string",
            "nullable": true
          },
          "companyNip": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Product": {
        "required": ["category"],
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "weight": {
            "type": "number",
            "format": "double"
          },
          "category": {
            "$ref": "#/components/schemas/ProductType"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "withdrawn": {
            "type": "boolean"
          },
          "shop": {
            "$ref": "#/components/schemas/ProductShop"
          }
        },
        "additionalProperties": false
      },
...
      "ProductShop": {
        "enum": [0, 1, 2],
        "type": "integer",
        "format": "int32"
      },
      "ProductType": {
        "enum": [0, 1, 2, 3, 4, 5, 6, 7],
        "type": "integer",
        "format": "int32"
      },
...
    }
  }
}